+++ docs/changelog.rst
@@ -11,6 +11,7 @@ CHANGELOG
 - Add attachments for OpenStreetMapParser
 - Add OpenStreetMap parser for Restricted Area
 - Add Apidae parser for Services
+- Add OpenStreetMap parser for infrastructures
 
 **Bug fixes**